requirements.txt
asgiref==3.9.2
diff-match-patch==20241021
Django==4.2.25
django-cors-headers==4.9.0
django-filter==25.1
django-import-export==4.3.10
djangorestframework==3.16.1
geojson==3.2.0
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.15.0
